Final working code.
Final version of the working code. (there are many features I would like to add, but it would require a lot of reconfiguring that I am unable to do and unwilling to pay for)
To switch modes just power cycle the device.
The edit Strokes, Timers, and Distance you will have to edit the file and reupload. The editable values are located at the beginning of the code under "session settings"